Simplify RDF loading from URL
- Renames RdfFetcher -> RdfLoader to align name with functionality
- Removes possibility of exception: returns blank graph, just as it
  does in the cases of an empty file or web page
- Removes rails logging on fetch error since any number of situations
  can cause us not to have a graph
