--- app/services/rdf_fetcher.rb
@@ -1,6 +1,0 @@
-# Retreives a remote RDF file and returns its graph
-class RdfFetcher
-  def self.call(url)
-    RDF::Graph.load(url)
-  end
-end
